frontend: renamed login pages to be specific to github
