+++ frontend/src/components/ghLoginCallbackPage.js
