+++ frontend/src/components/index.js
@@ -1,22 +1,22 @@
 import AppHeader from "./appHeader"
 
 import HomePage from "./homePage"
-import LoginPage from "./loginPage"
-import LoginCallbackPage from "./loginCallbackPage"
+import GHLoginPage from "./ghLoginPage"
+import GHLoginCallbackPage from "./ghLoginCallbackPage"
 
 // All components which can be used in Vue templates
 export const components = {
 	AppHeader,
 
 	HomePage,
-	LoginPage,
-	LoginCallbackPage
+	GHLoginPage,
+	GHLoginCallbackPage
 }
 
 // Page routes
 export const HomePageRoute = "/"
-export const LoginPageRoute = "/login"
-export const LoginCallbackPageRoute = "/auth/github"
+export const GHLoginPageRoute = "/login"
+export const GHLoginCallbackPageRoute = "/auth/github"
 
 export const routes = [
 	{
@@ -24,11 +24,11 @@ export const routes = [
 		component: HomePage
 	},
 	{ 
-		path: LoginPageRoute,
-		component: LoginPage,
+		path: GHLoginPageRoute,
+		component: GHLoginPage,
 	},
 	{
-		path: LoginCallbackPageRoute,
-		component: LoginCallbackPage,
+		path: GHLoginCallbackPageRoute,
+		component: GHLoginCallbackPage,
 	}
 ]
